Move networkStatus, networkStake to networkProvider package.
