+++ src/_docs.ts
@@ -8,6 +8,5 @@ export * from "./address";
 export * from "./nonce";
 export * from "./balance";
 export * from "./networkConfig";
-export * from "./networkStake";
 export * from "./networkParams";
 export * from "./utils";
