+++ src/index.ts
@@ -15,7 +15,6 @@ export * from "./boundaryAdapters";
 export * from "./interactive";
 export * from "./logger";
 export * from "./networkConfig";
-export * from "./networkStake";
 export * from "./networkParams";
 export * from "./signableMessage";
 export * from "./utils";
