+++ src/networkProvider/interface.ts
@@ -1,12 +1,12 @@
 import { BigNumber } from "bignumber.js";
 import { AccountOnNetwork } from "./accounts";
 import { NetworkConfig } from "../networkConfig";
-import { NetworkStake } from "../networkStake";
-import { NetworkStatus } from "../networkStatus";
+import { NetworkStake } from "./networkStake";
 import { Query } from "../smartcontracts";
 import { Stats } from "../stats";
 import { TransactionOnNetwork } from "./transactions";
 import { TransactionStatus } from "./transactionStatus";
+import { NetworkStatus } from "./networkStatus";
 
 /**
  * An interface that defines the endpoints of an HTTP API Provider.
