+++ src/networkProvider/proxyNetworkProvider.ts
@@ -2,8 +2,7 @@ import axios, { AxiosRequestConfig } from "axios";
 import { AccountOnNetwork } from "./accounts";
 import { IAddress, IContractQueryResponse, IDefinitionOfFungibleTokenOnNetwork, IDefinitionOfTokenCollectionOnNetwork, IFungibleTokenOfAccountOnNetwork, IHash, INetworkProvider, INonce, INonFungibleTokenOfAccountOnNetwork, ITransaction, Pagination } from "./interface";
 import { NetworkConfig } from "../networkConfig";
-import { NetworkStake } from "../networkStake";
-import { NetworkStatus } from "../networkStatus";
+import { NetworkStake } from "./networkStake";
 import { Query } from "../smartcontracts";
 import { Stats } from "../stats";
 import { ContractQueryResponse } from "./contractResults";
@@ -13,6 +12,7 @@ import { TransactionStatus } from "./transactionStatus";
 import { Hash } from "./primitives";
 import { ErrNetworkProvider } from "./errors";
 import { defaultAxiosConfig } from "./config";
+import { NetworkStatus } from "./networkStatus";
 
 // TODO: Find & remove duplicate code between "ProxyNetworkProvider" and "ApiNetworkProvider".
 export class ProxyNetworkProvider implements INetworkProvider {
